requirements.txt
langchain==0.1.11
python-dotenv==1.0.1
streamlit==1.31.1
